pyproject.toml
[tool.poetry]
name = "machine-learning-py"
version = "0.1.0"
description = "A machine learning using Poetry"
authors = ["Viktar Duzh <Viktar.Duzh@gmail.com>"]

packages = [
    { include = "apps/*" },
    { include = "libs/*" },
]

[tool.poetry.scripts]
speach-to-text = "apps.speach_to_text:main"

[tool.poetry.dependencies]
#python = "^3.11"
python = ">=3.11,<3.12"

# Shared libraries

# Application-specific dependencies
openai = "^1.12.0"
pychatgpt = "^0.4.3.3"
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.5"
chromadb = "^0.4.22"
unstructured = "^0.12.4"
pypdf = "^4.0.1"
rapidocr-onnxruntime = "^1.3.11"
faiss-cpu = "^1.7.4"
docarray = "^0.40.0"
langchainhub = "^0.1.14"
duckduckgo-search = "^4.4.3"
langchain = "^0.1.8"
langchain-community = "^0.0.21"
langchain-experimental = "^0.0.52"
google-api-python-client = ">=2.100.0"
